requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
blas=1.0=mkl
bottleneck=1.3.5=py38h080aedc_0
ca-certificates=2022.07.19=haa95532_0
certifi=2022.6.15=py38haa95532_0
cloudpickle=2.0.0=pyhd3eb1b0_0
cudatoolkit=11.0.221=h74a9793_0
cytoolz=0.11.0=py38he774522_0
dask-core=2022.7.0=py38haa95532_0
freetype=2.10.4=hd328e21_0
fsspec=2022.3.0=py38haa95532_0
hdf5=1.10.4=h7ebc959_0
icc_rt=2019.0.0=h0cc432a_1
imageio=2.19.3=py38haa95532_0
intel-openmp=2021.4.0=haa95532_3556
jpeg=9b=hb83a4c4_2
libopencv=4.0.1=hbb9e17c_0
libpng=1.6.37=h2a8f88b_0
libtiff=4.2.0=he0120a3_1
libuv=1.40.0=he774522_0
libwebp=1.2.2=h2bbff1b_0
locket=1.0.0=py38haa95532_0
lz4-c=1.9.3=h2bbff1b_1
mkl=2021.4.0=haa95532_640
mkl-service=2.4.0=py38h2bbff1b_0
mkl_fft=1.3.1=py38h277e83a_0
mkl_random=1.2.2=py38hf11a4ad_0
networkx=2.8.4=py38haa95532_0
ninja=1.10.2=haa95532_5
ninja-base=1.10.2=h6d14046_5
numexpr=2.8.3=py38hb80d3ca_0
numpy=1.22.3=py38h7a0a035_0
numpy-base=1.22.3=py38hca35cd5_0
opencv=4.0.1=py38h2a7c758_0
openssl=1.1.1q=h2bbff1b_0
packaging=21.3=pyhd3eb1b0_0
pandas=1.4.3=py38hd77b12b_0
partd=1.2.0=pyhd3eb1b0_1
pillow=9.2.0=py38hdc2b20a_1
pip=22.1.2=py38haa95532_0
py-opencv=4.0.1=py38he44ac1e_0
pyparsing=3.0.4=pyhd3eb1b0_0
python=3.8.13=h6244533_0
python-dateutil=2.8.2=pyhd3eb1b0_0
pytorch=1.7.1=py3.8_cuda110_cudnn8_0
pytz=2022.1=py38haa95532_0
pywavelets=1.3.0=py38h2bbff1b_0
pyyaml=6.0=py38h2bbff1b_1
scikit-image=0.19.2=py38hf11a4ad_0
scipy=1.7.3=py38h0a974cb_0
setuptools=61.2.0=py38haa95532_0
six=1.16.0=pyhd3eb1b0_1
sqlite=3.38.5=h2bbff1b_0
tifffile=2020.10.1=py38h8c2d366_2
tk=8.6.12=h2bbff1b_0
toolz=0.11.2=pyhd3eb1b0_0
torchaudio=0.7.2=py38
torchvision=0.8.2=py38_cu110
typing_extensions=4.1.1=pyh06a4308_0
vc=14.2=h21ff451_1
vs2015_runtime=14.27.29016=h5e58377_2
wheel=0.37.1=pyhd3eb1b0_0
wincertstore=0.2=py38haa95532_2
xz=5.2.5=h8cc25b3_1
yaml=0.2.5=he774522_0
zlib=1.2.12=h8cc25b3_2
zstd=1.5.2=h19a0ad4_0
